document torch.testing.assert_allclose
[ghstack-poisoned]
