+++ docs/source/testing.rst
@@ -6,3 +6,4 @@ torch.testing
 
 .. autofunction:: assert_close
 .. autofunction:: make_tensor
+.. autofunction:: assert_allclose
